refactor(canvas):  refactor Canvas component structure
